MCMC burnin uses a while loop at the sampler one step at a time, this eliminates the maximum number of steps to be set ahead of time and will just run however long is necessary to burn in, this makes the system much more automatic and removes running into the previous set limit when burn in takes longer
This might cause a problem with memory fragmentation due to reallocation of the data structures over time. Tests will be run on this but the impact should be minor.

Set stopping to 50 * tau_mult which is what the emcee paper suggests and based on tests it is working fine now that the stopping criteria is set correctly.