For MCMC with short simulations only 11% of the time was being spent running the simulations and the rest was being spent doing processing and writing the data
I have added a vectorized untransform which in one of the test cases reduced the function calls needed to convert the data from 6.6M to 7 which provides a substational speedup

I also added a checkpoint interval. If a generation takes only a few seconds to run it makes no sense to checkpoint at every generation so now the system will checkpoint by default every 5 minutes and the variable checkpointInterval can be set in the cache file to control this interval

corner plots in generate_graphs draw from 5 to 95 percentile. Outside this range MCMC accuracy degrades rapidly and it makes the bulk of the graph harder to read to include this range

removed all gzip compression from hdf5 files, it was not able to compress the data much anyways and slowed the process down