Fixed ordering of log_posterior_vectorize, since the change to async map also changed the order of results this was screwing up the MCMC process (this was an annoying bug to track down)
the mirror function is now MUCH better. It was incorrectly mirroring AbsoluteTime and AbsoluteHeight and making them symmetrical. The MCMC results should be more correct now (this also means previous results should be rerun since they did not correctly capture the non-symmetrical nature of some of the errors)
Bumped the version number (if everything works in testing this will be a new release)