Look at 90% of peak max and not peak max, looking at peak max (in the time dimension) effectively discretizes the answer to the simulation time points which is often 1/s and this then means the results is to the nearest second which leads to the result not being continuous which then degrades the kernel density estimator substantially by choosing 0.9 there are points above and below that value and a spline can be used to find the actual time where 0.9 is hit.
for kde_generator write out all the errors generated in the error_model to an h5 file. This is for debugging and makes it much easier to figure out where something went wrong

The pearson correlation runs from -1 to 1 and so far I have normalized this from 0 to 1 with a simple linear transform. However the negative correlation (-1 to 0) range looks like it is not possible to hit in chromatography and this is bunching up the values near the upper end of the range. Instead I have just chopped it off so that the system runs from 0 to 1 and any negative values are converted to 0