reference directions based on population size (may still need tuning) pearsonr_mat added in area to the score to test (gives more precise offsets but it may or may not help optimization) smooting uses a combo approach with bessel for smoothing the chromatogram and butter for smoothing the derivative
This has to do with a ringing effect. For the chromatogram a sharp change that happens in industrial data can cause a ringing effect if a butter filed is used by the bessel filter handles. However with the dervative a very sharp sharp transition like that is not observed so the ringing effect isn't there but some high frequency noise from the dervative leaks through

bumped version number